Upgrade replit/content
